Set blueprint on result of function if applicable
All Blueprints apply inheritance information to the object they modify.
Most Blueprints operate on an already existing object instance and for
them it's fine to apply this information to `this` because `this` refers
to the instance itself.

This commit allows Blueprints to do stuff like create a new class from
`this` by using Object.create(this). It's important to change only the
inheritance chain in the resulting class, and not in the original
class - since this kind of modification (store `this` as proto, add
more) is specific for producing new 'classes' via a Blueprint. To make
use of this, the function supplied when creating the Blueprint must
return the new class. If the original `this` or nothing is returned, the
original class/instance will be modified, exactly as before.