Fix reference resolving
Previously, the user-supplied resolver function did the work of both
getting and (optionally) processing the retrieved value. This
processing was prone to failure if the value retrieved was itself
going to be replaced by a reference (because the processing function
would be expecting a different type of value).

This commit adds the requirement of a 'getter' method used solely to
retrieve the value of interest. Once that value is found (call it X),
it is checked to see if it is a concrete value or a value that will
resolve to a reference (call it R) in the future. If that value needs
resolving, the value which depends on it will not receive the
placeholder value X. Instead, when the real value R is available, that's
what will be sent to the processessing function.